add security delete file

+++ backend/src/resolvers/file.resolver.ts
@@ -1,7 +1,7 @@
 import {
   Arg,
   Authorized,
-  Float,
+  Ctx,
   Mutation,
   Query,
   Resolver,
@@ -13,6 +13,8 @@ import {
 } from "../entities/file.entity";
 import FilesService from "../services/files.service";
 import { Message } from "../entities/user.entity";
+import { MyContext } from "..";
+import UserProjectAccessesService from "../services/userProjectAccesses.service";
 
 @Resolver()
 export class FileResolver {
@@ -68,7 +70,24 @@ export class FileResolver {
 
   @Authorized()
   @Mutation(() => Message)
-  async deleteFile(@Arg("id") id: number) {
+  async deleteFile(@Arg("id") id: number, @Ctx() context: MyContext) {
+
+    if (context.user == null)
+      throw new Error("Access denied! You need to be authenticated to perform this action!");
+    
+    const fileById = await new FilesService().findById(id);
+    if (!fileById) throw new Error("File does not exit");
+
+    const listUsersAccessesProject = await new UserProjectAccessesService().findUsersByAccessesProject(+fileById.project.id);
+
+    const findUserRoleAccessesProject = listUsersAccessesProject.find(user => user.user_id === context.user?.id);
+
+    if (!findUserRoleAccessesProject)
+      throw new Error("You do not have access to this project!");
+
+    if (findUserRoleAccessesProject.role === "VIEWER" && context.user.role !== "ADMIN")
+      throw new Error("You must be an owner or editor to delete this file!");
+    
     const delFile = await new FilesService().delete(id);
     const m = new Message();
     if (delFile) {
